* Split music off from sound_game into its own file (at least for FMOD)
