+++ src/engine/audio/CMakeLists.txt
@@ -4,6 +4,7 @@ list(APPEND GAME_SOURCES
 	"${CMAKE_CURRENT_SOURCE_DIR}/init.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/sound.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/sound_game.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/music.cpp"
 )
 list(APPEND EDITOR_SOURCES
 	"${CMAKE_CURRENT_SOURCE_DIR}/defines.cpp"
